+++ R/disabled.R
@@ -32,7 +32,7 @@
 #' disabled(numericInput("num", NULL, 5), dateInput("date", NULL))
 #' @export
 disabled <- function(...) {
-  tags <- list(...)
+  tags <- rlang::list2(...)
 
   # recursively add the disabled class to all tags
   if (length(tags) == 1 && inherits(tags[[1]], "shiny.tag")) {